requirements.txt
pandas==1.5.3
Pillow==10.1.0
python-barcode==0.15.1